Moved update logic from symbols to feature layer
This refactoring simplifies symbol logic and lays the groundwork for feature editing.

The symbols are no longer responsible for keeping track of primitive ids or to call update on primitives. Instead, a `render` call on a symbol now returns a set of `RenderPrimitive`s, which are then added to a render bundle by the feature layer.