Make vector tile layer init synchronous (#104)
There are quite a few issues with having the layer initialization method
asynchronous. This was originally done to allow sending vector tile
styles to the web workers when we initialize the layer, but I think we
must find a better approach to do that.

So this commit changes the initialization API to sync. No changes to the
tile processor are needed as sending the styles to the workers is not
yet implemented.